build: Switch from zipfile to git checkout
Having a part of the build on GitLab runners and the other parts in snapcraft is not optimal. Since we  cut down build time significantly with the precompiled ruby snap, fetching the gems shouldn’t be too much overhead, and prevents timing errors when the GitLab artifacts were deleted after one month.

Signed-off-by: Tobias Grasse <tg@glancr.de>