fix: Remove unsupported MySQL configuration parameter
MySQL 8 removed the `innodb_large_prefix` configuration parameter and will throw an error when my.cnf has this line. Existing installations were initialized with this parameter, so we check for its existence in the post-refresh hook and remove it if necessary. Also removed it from the initialized configuration file on new installations.

Signed-off-by: Tobias Grasse <tg@glancr.de>